refactor: useProduct, useRandom 의존성 분리
